Issue 83: Fixing some paths errors during build process.
