+++ gradle.properties
@@ -15,17 +15,13 @@ dcosAddress=master.mesos
 pravegaVersion=0.3.0-50.f56b52d-SNAPSHOT
 connectorVersion=0.2.1
 
-### outputs
+### Pravega-samples output library
 samplesVersion=0.3.0-SNAPSHOT
 
-### Flink
+### Flink-connector examples
 flinkVersion=1.4.0
 
-### Hadoop
+### Hadoop-connector examples
 hadoopVersion=2.8.1
-
-### Scala
 scalaVersion=2.11.8
-
-### Spark
 sparkVersion=2.2.0
