+++ pravega-client-examples/src/main/java/io/pravega/example/consolerw/Constants.java
@@ -8,7 +8,7 @@
  *   http://www.apache.org/licenses/LICENSE-2.0
  *   
  */
-package example.gettingstarted;
+package io.pravega.example.consolerw;
 
 /**
  * Defines a handful of constants shared by classes in this package.
@@ -16,9 +16,6 @@ package example.gettingstarted;
  */
 public class Constants {
     protected static final String DEFAULT_SCOPE = "examples";
-    protected static final String DEFAULT_STREAM_NAME = "helloStream";
+    protected static final String DEFAULT_STREAM_NAME = "someStream";
     protected static final String DEFAULT_CONTROLLER_URI = "tcp://127.0.0.1:9090";
-    
-    protected static final String DEFAULT_ROUTING_KEY = "helloRoutingKey";
-    protected static final String DEFAULT_MESSAGE = "hello world";
 }
