+++ pravega-client-examples/src/main/java/io/pravega/example/gettingstarted/Constants.java
@@ -8,12 +8,17 @@
  *   http://www.apache.org/licenses/LICENSE-2.0
  *   
  */
-package example.streamcuts;
+package io.pravega.example.gettingstarted;
 
 /**
  * Defines a handful of constants shared by classes in this package.
+ *
  */
 public class Constants {
-    static final String DEFAULT_SCOPE = "streamCutsExamples";
-    static final String DEFAULT_CONTROLLER_URI = "tcp://127.0.0.1:9090";
+    protected static final String DEFAULT_SCOPE = "examples";
+    protected static final String DEFAULT_STREAM_NAME = "helloStream";
+    protected static final String DEFAULT_CONTROLLER_URI = "tcp://127.0.0.1:9090";
+    
+    protected static final String DEFAULT_ROUTING_KEY = "helloRoutingKey";
+    protected static final String DEFAULT_MESSAGE = "hello world";
 }
