+++ pravega-client-examples/src/main/java/io/pravega/example/gettingstarted/HelloWorldWriter.java
@@ -8,22 +8,17 @@
  *   http://www.apache.org/licenses/LICENSE-2.0
  *   
  */
-package example.gettingstarted;
+package io.pravega.example.gettingstarted;
 
 import io.pravega.client.ClientFactory;
-import io.pravega.client.admin.ReaderGroupManager;
 import io.pravega.client.admin.StreamManager;
-import io.pravega.client.stream.EventRead;
-import io.pravega.client.stream.EventStreamReader;
-import io.pravega.client.stream.ReaderConfig;
-import io.pravega.client.stream.ReaderGroupConfig;
-import io.pravega.client.stream.ReinitializationRequiredException;
+import io.pravega.client.stream.EventStreamWriter;
+import io.pravega.client.stream.EventWriterConfig;
 import io.pravega.client.stream.ScalingPolicy;
-import io.pravega.client.stream.Stream;
 import io.pravega.client.stream.StreamConfiguration;
 import io.pravega.client.stream.impl.JavaSerializer;
 import java.net.URI;
-import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.DefaultParser;
@@ -32,54 +27,37 @@ import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 
 /**
- * A simple example app that uses a Pravega Reader to read from a given scope and stream.
+ * A simple example app that uses a Pravega Writer to write to a given scope and stream.
  */
-public class HelloWorldReader {
-    private static final int READER_TIMEOUT_MS = 2000;
+public class HelloWorldWriter {
 
     public final String scope;
     public final String streamName;
     public final URI controllerURI;
 
-    public HelloWorldReader(String scope, String streamName, URI controllerURI) {
+    public HelloWorldWriter(String scope, String streamName, URI controllerURI) {
         this.scope = scope;
         this.streamName = streamName;
         this.controllerURI = controllerURI;
     }
 
-    public void run() {
+    public void run(String routingKey, String message) {
         StreamManager streamManager = StreamManager.create(controllerURI);
         final boolean scopeIsNew = streamManager.createScope(scope);
+
         StreamConfiguration streamConfig = StreamConfiguration.builder()
                 .scalingPolicy(ScalingPolicy.fixed(1))
                 .build();
         final boolean streamIsNew = streamManager.createStream(scope, streamName, streamConfig);
 
-        final String readerGroup = UUID.randomUUID().toString().replace("-", "");
-        final ReaderGroupConfig readerGroupConfig = ReaderGroupConfig.builder().stream(Stream.of(scope, streamName)).build();
-        try (ReaderGroupManager readerGroupManager = ReaderGroupManager.withScope(scope, controllerURI)) {
-            readerGroupManager.createReaderGroup(readerGroup, readerGroupConfig);
-        }
-
         try (ClientFactory clientFactory = ClientFactory.withScope(scope, controllerURI);
-             EventStreamReader<String> reader = clientFactory.createReader("reader",
-                                                                           readerGroup,
-                                                                           new JavaSerializer<>(),
-                                                                           ReaderConfig.builder().build())) {
-            System.out.format("Reading all the events from %s/%s%n", scope, streamName);
-            EventRead<String> event = null;
-            do {
-                try {
-                    event = reader.readNextEvent(READER_TIMEOUT_MS);
-                    if (event.getEvent() != null) {
-                        System.out.format("Read event '%s'%n", event.getEvent());
-                    }
-                } catch (ReinitializationRequiredException e) {
-                    //There are certain circumstances where the reader needs to be reinitialized
-                    e.printStackTrace();
-                }
-            } while (event.getEvent() != null);
-            System.out.format("No more events from %s/%s%n", scope, streamName);
+             EventStreamWriter<String> writer = clientFactory.createEventWriter(streamName,
+                                                                                new JavaSerializer<>(),
+                                                                                EventWriterConfig.builder().build())) {
+            
+            System.out.format("Writing message: '%s' with routing-key: '%s' to stream '%s / %s'%n",
+                    message, routingKey, scope, streamName);
+            final CompletableFuture writeFuture = writer.writeEvent(routingKey, message);
         }
     }
 
@@ -91,17 +69,20 @@ public class HelloWorldReader {
         } catch (ParseException e) {
             System.out.format("%s.%n", e.getMessage());
             final HelpFormatter formatter = new HelpFormatter();
-            formatter.printHelp("HelloWorldReader", options);
+            formatter.printHelp("HelloWorldWriter", options);
             System.exit(1);
         }
-        
+
         final String scope = cmd.getOptionValue("scope") == null ? Constants.DEFAULT_SCOPE : cmd.getOptionValue("scope");
         final String streamName = cmd.getOptionValue("name") == null ? Constants.DEFAULT_STREAM_NAME : cmd.getOptionValue("name");
         final String uriString = cmd.getOptionValue("uri") == null ? Constants.DEFAULT_CONTROLLER_URI : cmd.getOptionValue("uri");
         final URI controllerURI = URI.create(uriString);
         
-        HelloWorldReader hwr = new HelloWorldReader(scope, streamName, controllerURI);
-        hwr.run();
+        HelloWorldWriter hww = new HelloWorldWriter(scope, streamName, controllerURI);
+        
+        final String routingKey = cmd.getOptionValue("routingKey") == null ? Constants.DEFAULT_ROUTING_KEY : cmd.getOptionValue("routingKey");
+        final String message = cmd.getOptionValue("message") == null ? Constants.DEFAULT_MESSAGE : cmd.getOptionValue("message");
+        hww.run(routingKey, message);
     }
 
     private static Options getOptions() {
@@ -109,11 +90,14 @@ public class HelloWorldReader {
         options.addOption("s", "scope", true, "The scope name of the stream to read from.");
         options.addOption("n", "name", true, "The name of the stream to read from.");
         options.addOption("u", "uri", true, "The URI to the controller in the form tcp://host:port");
+        options.addOption("r", "routingKey", true, "The routing key of the message to write.");
+        options.addOption("m", "message", true, "The message to write.");
         return options;
     }
 
     private static CommandLine parseCommandLineArgs(Options options, String[] args) throws ParseException {
         CommandLineParser parser = new DefaultParser();
-        return parser.parse(options, args);
+        CommandLine cmd = parser.parse(options, args);
+        return cmd;
     }
 }
